refactor: fix options, config and prompting (#15)
Properties are now initialised first to the current configuration values,
get extended by any options given to the constructor or parsed from
the arguments, and finally the user is prompted for properties that
are still undefined.  The configuration value, or a built-in default
value, is used as the default answer for any prompts that are issued.